requirements.txt
requests
vadersentiment
pandas
matplotlib
alpaca-py
flask
aiohttp
faker
